Update README.md
don't merge this. it has bugs


anyway, the Http request in checkNewAddress() should return something like this: (depending on the addressinputfields)
But where the heck do I check for this? for now the "add" button takes me back to the login-page. I don't know why. And I also don't know if the Httprequest for the openstreetmat api is still being sent.